make asyncmessage inner class
